0000471 nanbu
git-svn-id: https://plugins.svn.wordpress.org/usc-e-shop/trunk@571457 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ classes/usceshop.class.php
@@ -4912,7 +4912,7 @@ class usc_e_shop
 	
 	function getItemShippingCharge($post_id) {
 		$str = get_post_meta($post_id, '_itemShippingCharge', true);
-		return (int)$str;
+		return (float)$str;
 	}
 	
 	function getItemDeliveryMethod($post_id) {
